refactor GlobalEventFilters to EventFilters;
