+++ src/components/ConnectedDateFilterDialog.test.js
@@ -6,9 +6,6 @@ import { DateTime } from "luxon";
 import ConnectedDateFilterDialog from "./ConnectedDateFilterDialog";
 
 const initialState = {
-  globalFilters: {
-    hideEventsBefore: DateTime.fromISO("2018-07-07T00:00:00+01:00")
-  },
   events: {
     entries: [],
     assets: [],
@@ -16,6 +13,7 @@ const initialState = {
     refreshing: false
   },
   eventFilters: {
+    hideEventsBefore: DateTime.fromISO("2018-07-07T00:00:00+01:00"),
     selectedFilters: {
       categories: new Set(),
       date: null,
