+++ src/screens/CategoriesFilterScreen/index.test.js
@@ -12,9 +12,6 @@ const navigation: NavigationScreenProp<*> = ({ goBack: jest.fn() }: any);
 const mockStore = configureStore([thunk]);
 
 const initialState = {
-  globalFilters: {
-    hideEventsBefore: DateTime.fromISO("2018-07-07T00:00:00+01:00")
-  },
   events: {
     entries: [],
     assets: [],
@@ -22,6 +19,7 @@ const initialState = {
     refreshing: false
   },
   eventFilters: {
+    hideEventsBefore: DateTime.fromISO("2018-07-07T00:00:00+01:00"),
     selectedFilters: {
       categories: new Set(),
       date: null,
